Controller now works on the editor i guess
